Various changes to various versions.

+++ Groovy/src/main/groovy/meanStdDev_parallel.groovy
@@ -29,7 +29,7 @@ static Tuple meanStdDev(final Iterable<Number> data) {
       case 1: returnValue = new Tuple (r[0] / n, Double.NaN, 0); break
       default:
         final double xb = r[0] / n
-        returnValue = new Tuple (xb, sqrt(r[1] - n * xb * xb) / ((double) (n - 1)), n - 1)
+        returnValue = new Tuple (xb, sqrt((r[1] - n * xb * xb) / (n - 1)), n - 1)
         break
     }
   }
